16/2 update

+++ src/main/java/com/example/CorporateLearningManagmentSystem/dto/CourseDetails.java
@@ -7,16 +7,16 @@ import java.util.List;
 
 
 public class CourseDetails {
-    private Long courseId;
+    private int courseId;
     private String name;
     private String description;
     private int duration;
 
-    public Long getCourseId() {
+    public int getCourseId() {
         return courseId;
     }
 
-    public void setCourseId(Long courseId) {
+    public void setCourseId(int courseId) {
         this.courseId = courseId;
     }
 
@@ -44,7 +44,7 @@ public class CourseDetails {
         this.duration = duration;
     }
 
-    public CourseDetails(Long courseId, String name, String description, int duration) {
+    public CourseDetails(int courseId, String name, String description, int duration) {
         this.courseId = courseId;
         this.name = name;
         this.description = description;
