+++ src/main/java/com/example/CorporateLearningManagmentSystem/dto/UserRoleMapping.java
@@ -6,8 +6,8 @@ import lombok.Data;
 @Data
 public class UserRoleMapping {
 
-    private Long userId;
+    private int userId;
 
-    private Long rid;
+    private int rid;
 
 }
